chore(ocw): move filecoindot-ocw to filecoindot
