+++ Cargo.toml
@@ -1,7 +1,6 @@
 [workspace]
 members = [
     "filecoindot",
-    "filecoindot-ocw",
     "filecoindot-rpc",
 
     # node-example
