--- filecoindot-ocw/Cargo.toml
@@ -1,51 +1,0 @@
-[package]
-name = "filecoindot-ocw"
-authors = ['clearloop@chainsafe.io']
-description = 'filecoindot offchain worker apis.'
-version = "0.0.1"
-license = "GPL-3.0"
-homepage = 'https://github.com/ChainSafe/filecoindot'
-repository = 'https://github.com/ChainSafe/filecoindot'
-edition = '2018'
-
-[dependencies]
-hex = { version = "^0.4", features = [ "alloc" ], default-features = false }
-codec = { package = "parity-scale-codec", version = "2.3.1", default-features = false }
-serde = { version = "1.0.130", features = [ "derive" ], default-features = false }
-serde_json = { version = "1.0", features = [ "alloc" ], default-features = false }
-derive_more = "^0.99"
-
-# substrate dependencies
-sp-io = { git = 'https://github.com/paritytech/substrate', branch = 'polkadot-v0.9.10', default-features = false }
-frame-support = { git = 'https://github.com/paritytech/substrate', branch = 'polkadot-v0.9.10', default-features = false }
-frame-system = { git = 'https://github.com/paritytech/substrate', branch = 'polkadot-v0.9.10', default-features = false }
-sp-core = { git = 'https://github.com/paritytech/substrate', branch = 'polkadot-v0.9.10', default-features = false }
-
-# filecoindot
-filecoindot = { path = "../filecoindot", default-features = false }
-
-[dev-dependencies]
-bincode = "1.3.3"
-serde = { version = "1.0.130", features = [ "derive" ], default-features = false }
-serde_json = { version = "1.0", features = [ "alloc" ], default-features = false }
-reqwest = { version = "0.11.6", features = [ "json" ] }
-parking_lot = "0.11"
-futures = "0.3"
-tokio = { version = "1.13", features = [ "full" ] }
-thiserror = "1.0.30"
-dirs = "4.0.0"
-
-[features]
-default = [ "std" ]
-std = [
-  "hex/std",
-  "codec/std",
-  "serde/std",
-  "serde_json/std",
-  "sp-io/std",
-  "sp-core/std",
-  "frame-support/std",
-  "frame-system/std",
-  "filecoindot/std"
-]
-
