+++ filecoindot/Cargo.toml
@@ -12,11 +12,17 @@ edition = '2018'
 targets = ['x86_64-unknown-linux-gnu']
 
 [dependencies]
-codec = { package = "parity-scale-codec", version = "2.3.1", features = ['derive'], default-features = false }
+codec = { package = "parity-scale-codec", version = "2.3.1", default-features = false }
+serde = { version = "1.0.130", features = [ "derive" ], default-features = false }
+serde_json = { version = "1.0", features = [ "alloc" ], default-features = false }
+derive_more = "^0.99"
 
+# substrate dependencies
 frame-benchmarking = { git = 'https://github.com/paritytech/substrate', branch = 'polkadot-v0.9.10', default-features = false, optional = true }
 frame-support = { git = 'https://github.com/paritytech/substrate', branch = 'polkadot-v0.9.10', default-features = false }
 frame-system = { git = 'https://github.com/paritytech/substrate', branch = 'polkadot-v0.9.10', default-features = false }
+sp-core = { git = 'https://github.com/paritytech/substrate', branch = 'polkadot-v0.9.10', default-features = false }
+
 
 [dev-dependencies]
 serde = "1.0.130"
