+++ filecoindot/src/ocw/mod.rs
@@ -1,15 +1,25 @@
 // Copyright 2021 ChainSafe Systems
 // SPDX-License-Identifier: LGPL-3.0-only
+#![cfg_attr(not(feature = "std"), no_std)]
+#![allow(dead_code)]
 
-//! offchain worker logics
 use crate::{
-    api::{Api, ChainHeight},
-    result::{Error, Result},
-    Config,
+    ocw::{
+        api::{Api, ChainHeight},
+        result::{Error, Result},
+    },
+    Call, Config,
 };
-use filecoindot::Call;
 use frame_support::{log, sp_runtime::offchain::storage::StorageValueRef, sp_std::vec::Vec};
 use frame_system::offchain::{SendSignedTransaction, Signer};
+
+mod api;
+mod de;
+mod result;
+mod types;
+
+#[cfg(test)]
+mod tests;
 
 /// the storage key of filecoin rpc endpoint
 pub const FILECOIN_RPC: &[u8] = b"FILECOIN_RPC";
