+++ substrate-node-example/runtime/Cargo.toml
@@ -47,7 +47,6 @@ hex-literal = { version = "0.3.3", optional = true }
 
 # filecoindot dependencies
 filecoindot = { path = "../../filecoindot",  default-features = false }
-filecoindot-ocw = { path = "../../filecoindot-ocw",  default-features = false }
 
 [build-dependencies]
 substrate-wasm-builder = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.10" }
@@ -83,7 +82,6 @@ std = [
 	"sp-version/std",
 
 	"filecoindot/std",
-    "filecoindot-ocw/std",
 ]
 runtime-benchmarks = [
 	'frame-benchmarking',
